# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Kleines Mypage API
  description: API developed for Kleines Mypage
  license:
    name: MIT License
    url: https://github.com/nozomu-y/kleines-mypage-api/blob/main/LICENSE
  version: 1.0.0
servers:
  - url: https://www.chorkleines.com/member/mypage/api/v1/
    description: Main server
  - url: http://localhost:8888/kleines-mypage-api/v1/
    description: Development server
paths:
  /auth:
    post: 
      tags: 
        - auth
      summary: Authenticate user
      description: Authenticate user and return JWT
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: 
                  type: string
                  format: email
                password:
                  type: string
                  format: password
            example:
              email: example@gmail.com
              password: password
      responses:
        '200':
          description: ""
          content: 
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    format: JWT
                  user_id:
                    type: integer
              example:
                token: JSON Web Token
                user_id: 0
  /profile:
    get:
      tags:
        - profile
      summary: Get user profiles of multiple users
      description: Get user profiles of multiple users
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items:
              type: string
              enum: 
                - PRESENT
                - ABSENT
                - RESIGNED
        - name: part
          in: query
          schema:
            type: string
            enum:
              - S
              - A
              - T
              - B
        - name: grade
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of Profile models
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Profile' 
                example:
                  - user_id: 1
                    first_name: 名
                    last_name: 姓
                    name_kana: カナ
                    grade: 20
                    part: S
                    part_full: Soprano
                    status: PRESENT
                  - user_id: 2
                    first_name: 名
                    last_name: 姓
                    name_kana: カナ
                    grade: 19
                    part: A
                    part_full: Alto
                    status: ABSENT
  /profile/me/:
    get:
      tags:
      - profile
      summary: Get my profile 
      description: Get my profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile model
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Profile'
                example:
                  user_id: 1
                  first_name: 名
                  last_name: 姓
                  name_kana: カナ
                  grade: 20
                  part: S
                  part_full: Soprano
                  status: PRESENT
  /profile/{user_id}:
    get:
      tags:
      - profile
      summary: Get user profile from user_id
      description: Get user profile from user_id
      security:
        - bearerAuth: []
      parameters: 
        - name: user_id
          in: path
          required: true
          schema: 
            type: integer
      responses:
        '200':
          description: Profile model
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Profile'
                example:
                  user_id: 1
                  first_name: 名
                  last_name: 姓
                  name_kana: カナ
                  grade: 20
                  part: S
                  part_full: Soprano
                  status: PRESENT
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Profile:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: integer
          example: 1
        first_name:
          type: string
          example: 名
        last_name:
          type: string
          example: 姓
        name_kana:
          type: string
          example: カナ
        grade:
          type: integer
          example: 20
        part:
          type: string
          example: S
        part_full:
          type: string
          example: Soprano
        status:
          type: string
          example: PRESENT
tags:
  - name: auth
    description: Authentication
  - name: profile
    description: Profiles
externalDocs:
  description: GitHub
  url: https://github.com/nozomu-y/kleines-mypage-api